+++ crates/red_knot/src/module.rs
@@ -1,18 +1,11 @@
-use std::collections::BTreeMap;
-use std::fmt::{Display, Formatter};
-use std::hash::Hash;
-use std::num::{NonZeroU16, NonZeroUsize};
-use std::ops::{Deref, RangeFrom, RangeInclusive};
+use std::fmt::Formatter;
+use std::ops::Deref;
 use std::path::{Path, PathBuf};
-use std::str::FromStr;
 use std::sync::atomic::AtomicU32;
 use std::sync::Arc;
 
 use dashmap::mapref::entry::Entry;
-use rustc_hash::FxHashMap;
 use smol_str::SmolStr;
-
-use ruff_python_stdlib::identifiers::is_identifier;
 
 use crate::db::{QueryResult, SemanticDb, SemanticJar};
 use crate::files::FileId;
@@ -112,7 +105,7 @@ impl Module {
 ///
 /// Always normalized to the absolute form
 /// (never a relative module name, i.e., never `.foo`).
-#[derive(Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct ModuleName(smol_str::SmolStr);
 
 impl ModuleName {
@@ -175,12 +168,6 @@ impl Deref for ModuleName {
 
     fn deref(&self) -> &Self::Target {
         self.as_str()
-    }
-}
-
-impl From<&str> for ModuleName {
-    fn from(value: &str) -> Self {
-        Self::new(value)
     }
 }
 
@@ -781,312 +768,11 @@ impl PackageKind {
     }
 }
 
-#[derive(Debug, PartialEq, Eq)]
-struct TypeshedVersionsParseError {
-    line_number: NonZeroU16,
-    reason: TypeshedVersionsParseErrorKind,
-}
-
-impl Display for TypeshedVersionsParseError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let TypeshedVersionsParseError {
-            line_number,
-            reason,
-        } = self;
-        write!(
-            f,
-            "Error while parsing line {line_number} of typeshed's VERSIONS file: {reason}"
-        )
-    }
-}
-
-impl std::error::Error for TypeshedVersionsParseError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        if let TypeshedVersionsParseErrorKind::IntegerParsingFailure { err, .. } = &self.reason {
-            Some(err)
-        } else {
-            None
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum TypeshedVersionsParseErrorKind {
-    TooManyLines(NonZeroUsize),
-    UnexpectedNumberOfColons,
-    InvalidModuleName(String),
-    UnexpectedNumberOfHyphens,
-    UnexpectedNumberOfPeriods(String),
-    IntegerParsingFailure {
-        version: String,
-        err: std::num::ParseIntError,
-    },
-}
-
-impl Display for TypeshedVersionsParseErrorKind {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::TooManyLines(num_lines) => write!(
-                f,
-                "File has too many lines ({num_lines}); maximum allowed is {}",
-                NonZeroU16::MAX
-            ),
-            Self::UnexpectedNumberOfColons => {
-                f.write_str("Expected every non-comment line to have exactly one colon")
-            }
-            Self::InvalidModuleName(name) => write!(
-                f,
-                "Expected all components of '{name}' to be valid Python identifiers"
-            ),
-            Self::UnexpectedNumberOfHyphens => {
-                f.write_str("Expected every non-comment line to have exactly one '-' character")
-            }
-            Self::UnexpectedNumberOfPeriods(format) => write!(
-                f,
-                "Expected all versions to be in the form {{MAJOR}}.{{MINOR}}; got '{format}'"
-            ),
-            Self::IntegerParsingFailure { version, err } => write!(
-                f,
-                "Failed to convert '{version}' to a pair of integers due to {err}",
-            ),
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct TypeshedVersions(FxHashMap<ModuleName, PyVersionRange>);
-
-#[allow(unused)]
-impl TypeshedVersions {
-    fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    fn contains_module(&self, module_name: impl Into<ModuleName>) -> bool {
-        self.0.contains_key(&module_name.into())
-    }
-
-    fn module_exists_on_version(
-        &self,
-        module: impl Into<ModuleName>,
-        version: impl Into<PyVersion>,
-    ) -> bool {
-        let version = version.into();
-        let mut module = Some(module.into());
-        while let Some(module_to_try) = module {
-            if let Some(range) = self.0.get(&module_to_try) {
-                return range.contains(version);
-            }
-            module = module_to_try.parent();
-        }
-        false
-    }
-}
-
-impl FromStr for TypeshedVersions {
-    type Err = TypeshedVersionsParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut map = FxHashMap::default();
-
-        for (line_index, line) in s.lines().enumerate() {
-            // humans expect line numbers to be 1-indexed
-            let line_number = NonZeroUsize::new(line_index.saturating_add(1)).unwrap();
-
-            let Ok(line_number) = NonZeroU16::try_from(line_number) else {
-                return Err(TypeshedVersionsParseError {
-                    line_number: NonZeroU16::MAX,
-                    reason: TypeshedVersionsParseErrorKind::TooManyLines(line_number),
-                });
-            };
-
-            let Some(content) = line.split('#').map(str::trim).next() else {
-                continue;
-            };
-            if content.is_empty() {
-                continue;
-            }
-
-            let mut parts = content.split(':').map(str::trim);
-            let (Some(module_name), Some(rest), None) = (parts.next(), parts.next(), parts.next())
-            else {
-                return Err(TypeshedVersionsParseError {
-                    line_number,
-                    reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfColons,
-                });
-            };
-
-            let module_name = ModuleName::new(module_name);
-            if !module_name.components().all(is_identifier) {
-                return Err(TypeshedVersionsParseError {
-                    line_number,
-                    reason: TypeshedVersionsParseErrorKind::InvalidModuleName(
-                        module_name.to_string(),
-                    ),
-                });
-            }
-
-            match PyVersionRange::from_str(rest) {
-                Ok(version) => map.insert(module_name, version),
-                Err(reason) => {
-                    return Err(TypeshedVersionsParseError {
-                        line_number,
-                        reason,
-                    })
-                }
-            };
-        }
-
-        Ok(Self(map))
-    }
-}
-
-impl std::fmt::Display for TypeshedVersions {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let sorted_items: BTreeMap<&ModuleName, &PyVersionRange> = self.0.iter().collect();
-        for (module_name, range) in sorted_items {
-            writeln!(f, "{module_name}-{range}")?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-enum PyVersionRange {
-    AvailableFrom(RangeFrom<PyVersion>),
-    AvailableWithin(RangeInclusive<PyVersion>),
-}
-
-impl PyVersionRange {
-    fn contains(&self, version: PyVersion) -> bool {
-        match self {
-            Self::AvailableFrom(inner) => inner.contains(&version),
-            Self::AvailableWithin(inner) => inner.contains(&version),
-        }
-    }
-}
-
-impl FromStr for PyVersionRange {
-    type Err = TypeshedVersionsParseErrorKind;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut parts = s.split('-').map(str::trim);
-        match (parts.next(), parts.next(), parts.next()) {
-            (Some(lower), Some(""), None) => Ok(Self::AvailableFrom((lower.parse()?)..)),
-            (Some(lower), Some(upper), None) => {
-                Ok(Self::AvailableWithin((lower.parse()?)..=(upper.parse()?)))
-            }
-            _ => Err(TypeshedVersionsParseErrorKind::UnexpectedNumberOfHyphens),
-        }
-    }
-}
-
-impl std::fmt::Display for PyVersionRange {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AvailableFrom(range_from) => write!(f, "{}-", range_from.start),
-            Self::AvailableWithin(range_inclusive) => {
-                write!(f, "{}-{}", range_inclusive.start(), range_inclusive.end())
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash)]
-struct PyVersion {
-    major: u8,
-    minor: u8,
-}
-
-impl FromStr for PyVersion {
-    type Err = TypeshedVersionsParseErrorKind;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut parts = s.split('.').map(str::trim);
-        let (Some(major), Some(minor), None) = (parts.next(), parts.next(), parts.next()) else {
-            return Err(TypeshedVersionsParseErrorKind::UnexpectedNumberOfPeriods(
-                s.to_string(),
-            ));
-        };
-        let major = match u8::from_str(major) {
-            Ok(major) => major,
-            Err(err) => {
-                return Err(TypeshedVersionsParseErrorKind::IntegerParsingFailure {
-                    version: s.to_string(),
-                    err,
-                })
-            }
-        };
-        let minor = match u8::from_str(minor) {
-            Ok(minor) => minor,
-            Err(err) => {
-                return Err(TypeshedVersionsParseErrorKind::IntegerParsingFailure {
-                    version: s.to_string(),
-                    err,
-                })
-            }
-        };
-        Ok(Self { major, minor })
-    }
-}
-
-impl std::fmt::Display for PyVersion {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let PyVersion { major, minor } = self;
-        write!(f, "{major}.{minor}")
-    }
-}
-
-// TODO: unify with the PythonVersion enum in the linter/formatter crates?
-#[allow(unused)]
-#[derive(Copy, Clone, Hash, Debug, PartialEq, Eq, PartialOrd, Ord, Default)]
-enum SupportedPyVersion {
-    Py37,
-    #[default]
-    Py38,
-    Py39,
-    Py310,
-    Py311,
-    Py312,
-    Py313,
-}
-
-impl From<SupportedPyVersion> for PyVersion {
-    fn from(value: SupportedPyVersion) -> Self {
-        match value {
-            SupportedPyVersion::Py37 => PyVersion { major: 3, minor: 7 },
-            SupportedPyVersion::Py38 => PyVersion { major: 3, minor: 8 },
-            SupportedPyVersion::Py39 => PyVersion { major: 3, minor: 9 },
-            SupportedPyVersion::Py310 => PyVersion {
-                major: 3,
-                minor: 10,
-            },
-            SupportedPyVersion::Py311 => PyVersion {
-                major: 3,
-                minor: 11,
-            },
-            SupportedPyVersion::Py312 => PyVersion {
-                major: 3,
-                minor: 12,
-            },
-            SupportedPyVersion::Py313 => PyVersion {
-                major: 3,
-                minor: 13,
-            },
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::io::{Cursor, Read};
-    use std::num::{IntErrorKind, NonZeroU16, NonZeroU32};
+    use std::num::NonZeroU32;
     use std::path::{Path, PathBuf};
-    use std::str::FromStr;
 
     use zip::ZipArchive;
 
@@ -1094,8 +780,7 @@ mod tests {
     use crate::db::SourceDb;
     use crate::module::{
         path_to_module, resolve_module, set_module_search_paths, ModuleKind, ModuleName,
-        ModuleResolutionInputs, SupportedPyVersion, TypeshedVersions, TypeshedVersionsParseError,
-        TypeshedVersionsParseErrorKind, TYPESHED_STDLIB_DIRECTORY,
+        ModuleResolutionInputs, TYPESHED_STDLIB_DIRECTORY,
     };
     use crate::semantic::Dependency;
 
@@ -1107,9 +792,6 @@ mod tests {
         custom_typeshed: PathBuf,
         site_packages: PathBuf,
     }
-
-    #[allow(unsafe_code)]
-    const ONE: NonZeroU16 = unsafe { NonZeroU16::new_unchecked(1) };
 
     fn create_resolver() -> std::io::Result<TestCase> {
         let temp_dir = tempfile::tempdir()?;
@@ -1264,146 +946,6 @@ mod tests {
 
         assert!(functools_module_stub_source.contains("def update_wrapper("));
         Ok(())
-    }
-
-    #[test]
-    fn typeshed_versions() {
-        let versions_data = include_str!(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/vendor/typeshed/stdlib/VERSIONS"
-        ));
-
-        let versions = TypeshedVersions::from_str(versions_data).unwrap();
-        assert!(versions.len() > 100);
-        assert!(versions.len() < 1000);
-
-        // (will start failing if the stdlib adds a `foo` module, but oh well)
-        assert!(!versions.contains_module("foo"));
-        assert!(versions.contains_module("asyncio"));
-        assert!(versions.module_exists_on_version("asyncio", SupportedPyVersion::Py310));
-
-        assert!(versions.contains_module("asyncio.staggered"));
-        assert!(versions.module_exists_on_version("asyncio.staggered", SupportedPyVersion::Py38));
-        assert!(!versions.module_exists_on_version("asyncio.staggered", SupportedPyVersion::Py37));
-
-        assert!(versions.contains_module("audioop"));
-        assert!(versions.module_exists_on_version("audioop", SupportedPyVersion::Py312));
-        assert!(!versions.module_exists_on_version("audioop", SupportedPyVersion::Py313));
-    }
-
-    #[test]
-    fn invalid_typeshed_versions_bad_colon_number() {
-        assert_eq!(
-            TypeshedVersions::from_str("foo 3.7"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfColons
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("foo:: 3.7"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfColons
-            })
-        );
-    }
-
-    #[test]
-    fn invalid_typeshed_versions_non_identifier_modules() {
-        assert_eq!(
-            TypeshedVersions::from_str("not!an!identifier!: 3.7"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::InvalidModuleName(
-                    "not!an!identifier!".to_string()
-                )
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("(also_not).(an_identifier): 3.7"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::InvalidModuleName(
-                    "(also_not).(an_identifier)".to_string()
-                )
-            })
-        );
-    }
-
-    #[test]
-    fn invalid_typeshed_versions_bad_hyphen_number() {
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 3.8"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfHyphens
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 3.8--"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfHyphens
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 3.8--3.9"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfHyphens
-            })
-        );
-    }
-
-    #[test]
-    fn invalid_typeshed_versions_bad_period_number() {
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 38-"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfPeriods("38".to_string())
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 3..8-"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfPeriods(
-                    "3..8".to_string()
-                )
-            })
-        );
-        assert_eq!(
-            TypeshedVersions::from_str("foo: 3.8-3..11"),
-            Err(TypeshedVersionsParseError {
-                line_number: ONE,
-                reason: TypeshedVersionsParseErrorKind::UnexpectedNumberOfPeriods(
-                    "3..11".to_string()
-                )
-            })
-        );
-    }
-
-    #[test]
-    fn invalid_typeshed_versions_non_digits() {
-        let err = TypeshedVersions::from_str("foo: 1.two-").unwrap_err();
-        assert_eq!(err.line_number, ONE);
-        let TypeshedVersionsParseErrorKind::IntegerParsingFailure { version, err } = err.reason
-        else {
-            panic!()
-        };
-        assert_eq!(version, "1.two".to_string());
-        assert_eq!(*err.kind(), IntErrorKind::InvalidDigit);
-
-        let err = TypeshedVersions::from_str("foo: 3.8-four.9").unwrap_err();
-        assert_eq!(err.line_number, ONE);
-        let TypeshedVersionsParseErrorKind::IntegerParsingFailure { version, err } = err.reason
-        else {
-            panic!()
-        };
-        assert_eq!(version, "four.9".to_string());
-        assert_eq!(*err.kind(), IntErrorKind::InvalidDigit);
     }
 
     #[test]
